LPS-98937 Move logic to JspCDefaultsPlugin
